+++ src/cmd/cast.cpp
@@ -167,4 +167,4 @@ void do_cast(CharData *ch, char *argument, int/* cmd*/, int /*subcmd*/) {
 	}
 }
 
-
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
