+++ src/cmd/cast.h
@@ -6,3 +6,5 @@ class CharData;
 void do_cast(CharData *ch, char *argument, int/* cmd*/, int /*subcmd*/);
 
 #endif //BYLINS_CAST_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
