+++ src/cmd/create.h
@@ -8,3 +8,5 @@ class CharData;
 void do_create(CharData *ch, char *argument, int/* cmd*/, int subcmd);
 
 #endif //BYLINS_CREATE_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
