+++ src/cmd/flee.h
@@ -8,5 +8,4 @@ void DoFlee(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
 
 #endif //BYLINS_FLEE_H
 
-
 // vim: ts=4 sw=4 tw=0 noet syntax=cpp :
