+++ src/game_fight/fight_stuff.cpp
@@ -27,7 +27,7 @@
 #include "structs/global_objects.h"
 
 // extern
-void perform_drop_gold(CharData *ch, int amount);
+void PerformDropGold(CharData *ch, int amount);
 long GetExpUntilNextLvl(CharData *ch, int level);
 int max_exp_gain_pc(CharData *ch);
 int max_exp_loss_pc(CharData *ch);
@@ -669,7 +669,7 @@ void real_kill(CharData *ch, CharData *killer) {
 			obj_load_on_death(corpse, ch);
 		}
 		if (MOB_FLAGGED(ch, EMobFlag::kCorpse)) {
-			perform_drop_gold(ch, local_gold);
+			PerformDropGold(ch, local_gold);
 			ch->set_gold(0);
 		}
 		dl_load_obj(corpse, ch, nullptr, DL_ORDINARY);
