+++ src/game_skills/campfire.h
@@ -1,0 +1,10 @@
+#ifndef BYLINS_SRC_GAME_SKILLS_CAMPFIRE_H_
+#define BYLINS_SRC_GAME_SKILLS_CAMPFIRE_H_
+
+class CHarData;
+
+void DoCampfire(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_GAME_SKILLS_CAMPFIRE_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
