+++ src/game_skills/firstaid.h
@@ -1,0 +1,10 @@
+#ifndef BYLINS_SRC_GAME_SKILLS_FIRSTAID_H_
+#define BYLINS_SRC_GAME_SKILLS_FIRSTAID_H_
+
+class CharData;
+
+void DoFirstaid(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_GAME_SKILLS_FIRSTAID_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
